refactor(execution): add separate entrypoints file
