+++ src/execution/__tests__/abstract-test.ts
@@ -17,7 +17,7 @@ import { GraphQLSchema } from '../../type/schema.js';
 
 import { buildSchema } from '../../utilities/buildASTSchema.js';
 
-import { execute, executeSync } from '../execute.js';
+import { execute, executeSync } from '../entrypoints.js';
 
 interface Context {
   async: boolean;
