+++ src/execution/__tests__/schema-test.ts
@@ -16,7 +16,7 @@ import {
 } from '../../type/scalars.js';
 import { GraphQLSchema } from '../../type/schema.js';
 
-import { executeSync } from '../execute.js';
+import { executeSync } from '../entrypoints.js';
 
 describe('Execute: Handles execution with a complex schema', () => {
   it('executes using a schema', () => {
