+++ src/execution/__tests__/variables-test.ts
@@ -32,7 +32,10 @@ import { GraphQLSchema } from '../../type/schema.js';
 
 import { valueFromASTUntyped } from '../../utilities/valueFromASTUntyped.js';
 
-import { executeSync, experimentalExecuteIncrementally } from '../execute.js';
+import {
+  executeSync,
+  experimentalExecuteIncrementally,
+} from '../entrypoints.js';
 import { getVariableValues } from '../values.js';
 
 const TestFaultyScalarGraphQLError = new GraphQLError(
